feat: 实现团队成员管理 CRUD，重构 RBAC 角色枚举与租户包结构
- feat(tenant): 实现 EnterpriseTeamMembers 4 个接口（列表/邀请/改角色/移除）
  - 列表：2 表 JOIN 查询 + 分离 roleCode 查询（Map 合并），支持 sort/order 排序
  - 邀请：校验 enterpriseId → 校验角色域 → 防重复 → 保存（DataIntegrityViolationException）
  - 改角色：校验成员归属 → 校验 OWNER → 校验角色域 → UPDATE
  - 移除：查询成员 → 校验归属 → 校验 OWNER/非自己 → 逻辑删除
- refactor(tenant): 包结构调整，controller/ service/ 子包拆分，删除 TeamMemberUpdateVO
- feat(rbac): Role 枚举新增 roleScope 字段，新建 RoleScope 枚举（PLATFORM/ENTERPRISE/USER）
- feat(common): 新建 SelectOrder 枚举；ErrorCode 新增成员/角色域相关错误码
- fix(auth): SysUser 添加 autoResultMap = true 修复 phone 加密字段反序列化
- fix(framework): 添加 mybatis-plus-jsqlparser 依赖（兼容 MyBatis-Plus 3.5.15）
- refactor(auth): EnterprisesServiceImpl 移除 ServletRequest 注入，改为 SecureActionContext 参数
- docs: 更新 API 文档团队成员列表 sort/order 参数及 POST/PUT 响应结构

+++ modules/system/src/main/java/interview/system/rbac/model/vo/RoleListItemVO.java
@@ -2,8 +2,6 @@ package interview.system.rbac.model.vo;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
-
-import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 
 /**
@@ -12,7 +10,7 @@ import java.time.OffsetDateTime;
  */
 @Builder
 public record RoleListItemVO(
-    Long id,
+    Integer id,
     String roleCode,
     String roleName,
     String roleScope,
